# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoWork API
  version: 0.1.0
  description: 'V1 design contract. Milestones 1–3 implement GET /health, GET /ready, POST /api/v1/auth/login and GET /api/v1/auth/me;
    business endpoints remain planned. Read docs/contracts.md for ownership, lifecycle
    and validation rules. 1 MiB JSON body limit; reject unknown/repeated query parameters
    and duplicate JSON keys. Global routing errors: 404 NOT_FOUND, 405 METHOD_NOT_ALLOWED.
    All errors follow Error schema and all responses include X-Request-ID.'
servers:
- url: http://localhost:8080
  description: Planned local server, not currently running
security:
- bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      x-implemented-milestone: 3
      tags:
      - Auth
      summary: login
      description: Uniform 401 for missing user, wrong password or inactive user.
        No public registration. HS256 access token, 900 seconds.
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              description: Sanitized incoming correlation ID or generated UUID.
              schema:
                type: string
                minLength: 1
                maxLength: 64
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResultResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      x-implemented-milestone: 3
      tags:
      - Auth
      summary: getCurrentUser
      description: Requires active authenticated user; no user:read permission required.
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              description: Sanitized incoming correlation ID or generated UUID.
              schema:
                type: string
                minLength: 1
                maxLength: 64
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/401'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        name: &id010
          type: string
          minLength: 1
          maxLength: 100
        email: &id011
          type: string
          format: email
          minLength: 3
          maxLength: 254
          description: Trimmed lowercase ASCII email, no display-name form.
        role: *id002
        is_active:
          type: boolean
        created_at: &id003
          type: string
          format: date-time
        updated_at: *id003
      required:
      - id
      - name
      - email
      - role
      - is_active
      - created_at
      - updated_at
...
    Login:
      type: object
      additionalProperties: false
      properties:
        email: *id011
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 72
          writeOnly: true
          description: Nonempty, at most 72 UTF-8 bytes; unchanged input.
      required:
      - email
      - password
    LoginResult:
      type: object
      additionalProperties: false
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum:
          - Bearer
        expires_in:
          type: integer
          enum:
          - 900
        user:
          $ref: '#/components/schemas/User'
      required:
      - access_token
      - token_type
      - expires_in
      - user
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties:
                type: string
          required:
          - code
          - message
        request_id:
          type: string
          minLength: 1
          maxLength: 64
      required:
      - error
      - request_id
...
    UserResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/User'
      required:
      - data
...
    LoginResultResponse:
      type: object
      additionalProperties: false
      properties:
        data:
          $ref: '#/components/schemas/LoginResult'
      required:
      - data
...
  responses:
    '400':
      description: INVALID_JSON
      headers:
        X-Request-ID:
          description: Sanitized incoming correlation ID or generated UUID.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    '401':
      description: UNAUTHENTICATED or INVALID_CREDENTIALS
      headers:
        X-Request-ID:
          description: Sanitized incoming correlation ID or generated UUID.
          schema:
            type: string
            minLength: 1
            maxLength: 64
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer challenge on protected endpoints.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    '409':
      description: Domain conflict; see contracts.md error catalogue
      headers:
        X-Request-ID:
          description: Sanitized incoming correlation ID or generated UUID.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    '413':
      description: PAYLOAD_TOO_LARGE
      headers:
        X-Request-ID:
          description: Sanitized incoming correlation ID or generated UUID.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    '415':
      description: UNSUPPORTED_MEDIA_TYPE
      headers:
        X-Request-ID:
          description: Sanitized incoming correlation ID or generated UUID.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    '422':
      description: VALIDATION_ERROR
      headers:
        X-Request-ID:
          description: Sanitized incoming correlation ID or generated UUID.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    '500':
      description: INTERNAL_ERROR
      headers:
        X-Request-ID:
          description: Sanitized incoming correlation ID or generated UUID.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    '503':
      description: SERVICE_UNAVAILABLE or WRITE_OUTCOME_UNKNOWN
      headers:
        X-Request-ID:
          description: Sanitized incoming correlation ID or generated UUID.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'